document-genie/backend/requirements.txt
fastapi
uvicorn
qdrant-client
sentence-transformers
PyMuPDF
python-multipart
aiofiles
pymongo
python-dotenv
passlib[bcrypt]
groq
pydantic